Commands: split SwiftPackageTool into multiple files
