+++ Sources/Commands/CMakeLists.txt
@@ -7,6 +7,26 @@
 # See http://swift.org/CONTRIBUTORS.txt for Swift project authors
 
 add_library(Commands
+  PackageTools/APIDiff.swift
+  PackageTools/ArchiveSource.swift
+  PackageTools/CompletionTool.swift
+  PackageTools/ComputeChecksum.swift
+  PackageTools/Config.swift
+  PackageTools/Describe.swift
+  PackageTools/DumpCommands.swift
+  PackageTools/EditCommands.swift
+  PackageTools/Format.swift
+  PackageTools/Init.swift
+  PackageTools/Learn.swift
+  PackageTools/PluginCommand.swift
+  PackageTools/ResetCommands.swift
+  PackageTools/Resolve.swift
+  PackageTools/ShowDependencies.swift
+  PackageTools/SwiftPackageCollectionsTool.swift
+  PackageTools/SwiftPackageRegistryTool.swift
+  PackageTools/SwiftPackageTool.swift
+  PackageTools/ToolsVersionCommand.swift
+  PackageTools/Update.swift
   Snippets/CardEvent.swift
   Snippets/Cards/SnippetCard.swift
   Snippets/Cards/SnippetGroupCard.swift
@@ -17,13 +37,12 @@ add_library(Commands
   SwiftBuildTool.swift
   SwiftPackageCollectionsTool.swift
   SwiftPackageRegistryTool.swift
-  SwiftPackageTool.swift
   SwiftRunTool.swift
   SwiftTestTool.swift
   ToolWorkspaceDelegate.swift
   Utilities/APIDigester.swift
   Utilities/DependenciesSerializer.swift
-  Utilities/Describe.swift
+  Utilities/DescribedPackage.swift
   Utilities/DOTManifestSerializer.swift
   Utilities/GenerateLinuxMain.swift
   Utilities/MultiRootSupport.swift
