+++ Sources/Commands/PackageTools/ArchiveSource.swift
@@ -1,0 +1,56 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift open source project
+//
+// Copyright (c) 2014-2022 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See http://swift.org/LICENSE.txt for license information
+// See http://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+
+import ArgumentParser
+import SourceControl
+import TSCBasic
+
+extension SwiftPackageTool {
+    struct ArchiveSource: SwiftCommand {
+        static let configuration = CommandConfiguration(
+            commandName: "archive-source",
+            abstract: "Create a source archive for the package"
+        )
+
+        @OptionGroup(_hiddenFromHelp: true)
+        var globalOptions: GlobalOptions
+
+        @Option(
+            name: [.short, .long],
+            help: "The absolute or relative path for the generated source archive"
+        )
+        var output: AbsolutePath?
+
+        func run(_ swiftTool: SwiftTool) throws {
+            let packageRoot = try globalOptions.locations.packageDirectory ?? swiftTool.getPackageRoot()
+            let repository = GitRepository(path: packageRoot)
+
+            let destination: AbsolutePath
+            if let output = output {
+                destination = output
+            } else {
+                let graph = try swiftTool.loadPackageGraph()
+                let packageName = graph.rootPackages[0].manifest.displayName // TODO: use identity instead?
+                destination = packageRoot.appending(component: "\(packageName).zip")
+            }
+
+            try repository.archive(to: destination)
+
+            if destination.isDescendantOfOrEqual(to: packageRoot) {
+                let relativePath = destination.relative(to: packageRoot)
+                print("Created \(relativePath.pathString)")
+            } else {
+                print("Created \(destination.pathString)")
+            }
+        }
+    }
+}
