+++ Sources/Commands/PackageTools/Describe.swift
@@ -1,0 +1,71 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift open source project
+//
+// Copyright (c) 2014-2022 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See http://swift.org/LICENSE.txt for license information
+// See http://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+
+import ArgumentParser
+import Foundation
+import PackageModel
+import TSCBasic
+
+extension SwiftPackageTool {
+    struct Describe: SwiftCommand {
+        static let configuration = CommandConfiguration(
+            abstract: "Describe the current package")
+        
+        @OptionGroup(_hiddenFromHelp: true)
+        var globalOptions: GlobalOptions
+        
+        @Option(help: "json | text")
+        var type: DescribeMode = .text
+        
+        func run(_ swiftTool: SwiftTool) throws {
+            let workspace = try swiftTool.getActiveWorkspace()
+            
+            guard let packagePath = try swiftTool.getWorkspaceRoot().packages.first else {
+                throw StringError("unknown package")
+            }
+            
+            let package = try tsc_await {
+                workspace.loadRootPackage(
+                    at: packagePath,
+                    observabilityScope: swiftTool.observabilityScope,
+                    completion: $0
+                )
+            }
+            
+            try self.describe(package, in: type)
+        }
+        
+        /// Emits a textual description of `package` to `stream`, in the format indicated by `mode`.
+        func describe(_ package: Package, in mode: DescribeMode) throws {
+            let desc = DescribedPackage(from: package)
+            let data: Data
+            switch mode {
+            case .json:
+                let encoder = JSONEncoder.makeWithDefaults()
+                encoder.keyEncodingStrategy = .convertToSnakeCase
+                data = try encoder.encode(desc)
+            case .text:
+                var encoder = PlainTextEncoder()
+                encoder.formattingOptions = [.prettyPrinted]
+                data = try encoder.encode(desc)
+            }
+            print(String(decoding: data, as: UTF8.self))
+        }
+        
+        enum DescribeMode: String, ExpressibleByArgument {
+            /// JSON format (guaranteed to be parsable and stable across time).
+            case json
+            /// Human readable format (not guaranteed to be parsable).
+            case text
+        }
+    }
+}
