+++ Sources/Commands/PackageTools/ResetCommands.swift
@@ -1,0 +1,51 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift open source project
+//
+// Copyright (c) 2014-2022 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See http://swift.org/LICENSE.txt for license information
+// See http://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+
+import ArgumentParser
+
+extension SwiftPackageTool {
+    struct Clean: SwiftCommand {
+        static let configuration = CommandConfiguration(
+            abstract: "Delete build artifacts")
+
+        @OptionGroup(_hiddenFromHelp: true)
+        var globalOptions: GlobalOptions
+
+        func run(_ swiftTool: SwiftTool) throws {
+            try swiftTool.getActiveWorkspace().clean(observabilityScope: swiftTool.observabilityScope)
+        }
+    }
+
+    struct PurgeCache: SwiftCommand {
+        static let configuration = CommandConfiguration(
+            abstract: "Purge the global repository cache.")
+
+        @OptionGroup(_hiddenFromHelp: true)
+        var globalOptions: GlobalOptions
+
+        func run(_ swiftTool: SwiftTool) throws {
+            try swiftTool.getActiveWorkspace().purgeCache(observabilityScope: swiftTool.observabilityScope)
+        }
+    }
+
+    struct Reset: SwiftCommand {
+        static let configuration = CommandConfiguration(
+            abstract: "Reset the complete cache/build directory")
+
+        @OptionGroup(_hiddenFromHelp: true)
+        var globalOptions: GlobalOptions
+
+        func run(_ swiftTool: SwiftTool) throws {
+            try swiftTool.getActiveWorkspace().reset(observabilityScope: swiftTool.observabilityScope)
+        }
+    }
+}
