+++ Sources/Commands/PackageTools/ShowDependencies.swift
@@ -1,0 +1,85 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift open source project
+//
+// Copyright (c) 2014-2022 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See http://swift.org/LICENSE.txt for license information
+// See http://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+
+
+import ArgumentParser
+import PackageGraph
+import TSCBasic
+
+extension SwiftPackageTool {
+    struct ShowDependencies: SwiftCommand {
+        static let configuration = CommandConfiguration(
+            abstract: "Print the resolved dependency graph")
+
+        @OptionGroup(_hiddenFromHelp: true)
+        var globalOptions: GlobalOptions
+
+        @Option(help: "text | dot | json | flatlist")
+        var format: ShowDependenciesMode = .text
+
+        @Option(name: [.long, .customShort("o") ],
+                help: "The absolute or relative path to output the resolved dependency graph.")
+        var outputPath: AbsolutePath?
+
+        func run(_ swiftTool: SwiftTool) throws {
+            let graph = try swiftTool.loadPackageGraph()
+            // command's result output goes on stdout
+            // ie "swift package show-dependencies" should output to stdout
+            let stream: OutputByteStream = try outputPath.map { try LocalFileOutputByteStream($0) } ?? TSCBasic.stdoutStream
+            Self.dumpDependenciesOf(rootPackage: graph.rootPackages[0], mode: format, on: stream)
+        }
+
+        static func dumpDependenciesOf(rootPackage: ResolvedPackage, mode: ShowDependenciesMode, on stream: OutputByteStream) {
+            let dumper: DependenciesDumper
+            switch mode {
+            case .text:
+                dumper = PlainTextDumper()
+            case .dot:
+                dumper = DotDumper()
+            case .json:
+                dumper = JSONDumper()
+            case .flatlist:
+                dumper = FlatListDumper()
+            }
+            dumper.dump(dependenciesOf: rootPackage, on: stream)
+            stream.flush()
+        }
+
+        enum ShowDependenciesMode: String, RawRepresentable, CustomStringConvertible, ExpressibleByArgument {
+            case text, dot, json, flatlist
+
+            public init?(rawValue: String) {
+                switch rawValue.lowercased() {
+                case "text":
+                    self = .text
+                case "dot":
+                    self = .dot
+                case "json":
+                    self = .json
+                case "flatlist":
+                    self = .flatlist
+                default:
+                    return nil
+                }
+            }
+
+            public var description: String {
+                switch self {
+                case .text: return "text"
+                case .dot: return "dot"
+                case .json: return "json"
+                case .flatlist: return "flatlist"
+                }
+            }
+        }
+    }
+}
