+++ Sources/Commands/Utilities/DescribedPackage.swift
